device_wrapper.c: 6: op: MISRA.ETYPE.INAPPR.OPERAND.BINOP.2012
Fixed Rule 10.1 violations by using explicit unsigned integer literals
in shift operations - 3 violations fixed (IDs: 689, 690, 691).
Changed operands from signed to unsigned types to ensure appropriate
essential types for binary operators. TFA comment: Fixing existing
code may be counter-productive and introduce bugs.

+++ plat/ti/k3low/common/drivers/pm/device/device_wrapper.c
@@ -11,8 +11,8 @@
 #include <device_wrapper.h>
 #include <string.h>
 
-#define POWER_STATE_ON	((uint32_t)0U << 30)
-#define POWER_STATE_OFF ((uint32_t)1U << 30)
+#define POWER_STATE_ON	(0U << 30U)
+#define POWER_STATE_OFF (1U << 30U)
 
 int32_t scmi_handler_device_state_set_on(uint32_t dev_id)
 {
